backend/requirements.txt
Flask>=2.2
Flask-JWT-Extended==4.6.0
openai==0.27.0
psycopg2-binary
python-dotenv==1.0.1
flask_cors==5.0.0
